The test case does not print logs on the screen anymore.
The test case was immediately broken as soon as we add a new
directory in $GOBO/library because of the assert on `count'.
Furthermore, there is no guarantee as far as I know that the
directory names will be traversed in a given order. And finally,
the test should also run from the official delivery, where there
is no CVS anymore. Modified the test case to try to address
these problems.


git-svn-id: file:///c/gobo/svn/gobo/trunk@3076 fc6a2fa8-d027-0410-966b-9b90f175b7b2